Refactor stuff to own modules.
